+++ advent-of-code-2024/src/lib.rs
@@ -1,4 +1,4 @@
 pub trait Solution {
-    fn part1(&self) -> i32;
-    fn part2(&self) -> i32;
+    fn part1(&self) -> i128;
+    fn part2(&self) -> i128;
 }
